--- modules/core/constructor.ts
@@ -1,1 +1,0 @@
-export type Constructor<T extends {} = {}, A extends any[] = any[]> = new (...args: A) => T;
