+++ modules/core/resolver/base-runner.resolver.ts
@@ -1,4 +1,4 @@
-import { Constructor } from "@core/constructor";
+import { Constructor } from "@core/types/constructor";
 
 export interface IRunnerResolverConfigBase<R extends Constructor> {
     runners: R[];
