+++ modules/core/types/constructor.ts
@@ -1,0 +1,4 @@
+import { JsonObject } from "./json-object";
+
+export type Constructor<T extends {} = {}, A extends any[] = any[]> = new (...args: A) => T;
+export type RunnerConstructor<T extends {} = {}, A extends JsonObject[] = JsonObject[]> = Constructor<T,A>;
