+++ modules/promise/dev/worker-runner.resolver.ts
@@ -1,6 +1,6 @@
 import { IWorkerCommand } from "@core/commands/worker-commands";
-import { Constructor } from "@core/constructor";
 import { WorkerRunnerResolverBase } from "@core/resolver/worker-runner.resolver";
+import { Constructor } from "@core/types/constructor";
 import { DevWorkerBridge } from "@core/worker-bridge/dev-worker-bridge";
 
 export class DevWorkerRunnerResolver<R extends Constructor<{[key: string]: any}>> extends WorkerRunnerResolverBase<R> {
