+++ modules/rxjs/node-runner.resolver.ts
@@ -1,0 +1,6 @@
+import { Constructor } from "@core/constructor";
+import { NodeRunnerResolverBase } from "@core/resolver/node-runner.resolver";
+
+export class NodeRunnerResolver<R extends Constructor> extends NodeRunnerResolverBase<R> {
+    
+}
