+++ modules/rxjs/worker-runner.resolver.ts
@@ -1,5 +1,5 @@
+import { Constructor } from "@core/constructor";
 import { WorkerRunnerResolverBase } from "@core/resolver/worker-runner.resolver";
-import { Constructor } from "@core/types/constructor";
 
 export class WorkerRunnerResolver<R extends Constructor<{[key: string]: any}>> extends WorkerRunnerResolverBase<R> {
