+++ package.json
@@ -25,8 +25,11 @@
     "karma-webpack": "^4.0.2",
     "ts-loader": "^6.2.1",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "typescript": "^3.7.2",
+    "typescript": "^3.7.3",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.10"
+  },
+  "dependencies": {
+    "rxjs": "^6.5.3"
   }
 }
